requirements.txt
gspread==5.7.2
oauth2client==4.1.3
mysql-connector-python==8.0.33